oamc: Change the street splitting min distance.
For streets longer than 900m (?) the tool would split them into
neighbourhoods assuming that those were different streets with colliding
names in the same city.  They would be displayed to the user as:
* City -> Street (Neighbourhood A)
* City -> Street (Neighbourhood B)
* etc
Users have complained that this is unpractical because often streets are
much longer and when they search for a housenumber, they usually only
know the city, the street name, and the number.

It was more problematic for address nodes with "addr:place" because
those places can sometimes span 20-25km, so OsmAnd would display them
as:

* Municipality -> Village (some smaller village A)
* Municipality -> Village (hamlet B, etc.)
* etc.
Now the minimum splitting distance is set to 5km (?) for streets and
60km (?) for places.